Deployment/requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
scikit-learn == 1.0.2